pyproject.toml
[tool.poetry]
name = "linktrove"
version = "0.1.0"
description = "A place to keep your links"
authors = ["Daniel Nieto <daniel@nieto.dev>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-allauth = "^0.61.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
djlint = "^1.34.1"
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.djlint]
profile="django"
